requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-filter==23.2
sqlparse==0.4.4
